% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smiles_query.R
\name{smiles_query}
\alias{smiles_query}
\title{Chemical Structure searches with SMILES or SMARTS strings}
\usage{
smiles_query(smiles, type = c("exact", "substructure", "superstructure",
  "similarity"), similarity = NA)
}
\arguments{
\item{smiles}{a character of smiles}

\item{type}{}
}
\value{
For \code{str_match}, a character matrix. First column is the
  complete match, followed by one column for each capture group.
  For \code{str_match_all}, a list of character matrices.
}
\description{
This query performs chemical structure searches using SMILES and SMARTS linear notations.
}
\examples{
structure_info <- smiles_query('OC(=O)c1ccc(OCc2ccccc2)cc1', type = 'similarity', similarity = 0.7)
}
\seealso{
\code{\link{str_extract}} to extract the complete match,
  \code{\link[stringi]{stri_match}} for the underlying
  implementation.
}

